tinker: Debug [LocalDateTime] serialization/deserialization issue
Attempted to store separate orders (currently named as class: [MenuItemInfoListWrapper]) instead of appending every order's [MenuItemInfo] onto a running list.

Each order has a field of type [LocalDateTime], used as a timestamp for creation.

It wasn't an issue between client-server apps converting from json to java objects. It was a local issue involving how [LocalDateTime] is an immutable type and define its fields with private access, but Gson serializing/deserializing required public access. The solution was to define a [TypeAdapter] and register it with [GsonBuilder] before converting the model containing a [LocalDateType] field into json.

ref: https://howtodoinjava.com/gson/gson-typeadapter-for-inaccessibleobjectexception/